+++ tests/test_misc.py
@@ -78,31 +78,7 @@ def test_plot_and_eval_ddfs(ddf_d, tmpdir, logger):
     func(ddf_d, log=logger, out_dir=tmpdir)
     
     
-#===============================================================================
-# file-based tests ------
-#===============================================================================
-from .data.bldgs_data_scripts import load_tests_cases_from_file, test_cases_l
-load_tests_cases_from_file() #setup file-based test cases
 
-
-
-@pytest.mark.dev
-@pytest.mark.parametrize('testCase',
-    #['case4_R2']
-     test_cases_l,    
-    )
-def test_bldgs_workflow(testCase, 
-                        ci_df, bldg_meta_d,fixed_costs_d,  #from conftest based on testCase                       
-                        tmpdir, logger):
-    """end-to-end backend testing"""
-    from misc.bldgs_script_example import bldgs_workflow as func
-    
-    func(ci_df, bldg_meta_d=bldg_meta_d, fixed_costs_d=fixed_costs_d,
-         #settings_d=settings_d, #use default 
-          curve_name=f'{testCase}_c00',
-         logger=logger, out_dir=tmpdir,
-         plot=False,
-         )
